*** empty log message ***

+++ bin/win32fe/petsctlibfe.cpp
@@ -1,6 +1,5 @@
-/* $Id: petsctlibfe.cpp,v 1.2 2001/03/23 19:31:16 buschelm Exp buschelm $ */
+/* $Id: petsctlibfe.cpp,v 1.3 2001/03/23 19:37:34 buschelm Exp $ */
 #include <fstream>
-#include <iostream>
 #include "Windows.h"
 #include "petsctlibfe.h"
 
@@ -9,19 +8,27 @@ using namespace std;
 
 void tlib::Execute() {
   archiver::Execute();
-  string temp;
-  temp = *file.begin();
-  temp = temp.substr(0,temp.rfind("."));
-  temp = temp.substr(1) + ".BAK";
-  bool deleteme = FALSE;
-  {
-    ifstream LibraryExists(temp.c_str());
-    if (LibraryExists) deleteme=TRUE;
+  if (!helpfound) {
+    string temp;
+    temp = *file.begin();
+    temp = temp.substr(0,temp.rfind("."));
+    temp = temp.substr(1) + ".BAK";
+    bool deleteme = FALSE;
+    { /* Open file stream */ 
+      ifstream LibraryExists(temp.c_str());
+      if (LibraryExists) deleteme=TRUE;
+    } /* Close file stream */
+    if (deleteme) {
+      if (verbose) cout << "del \"" << temp << "\"" << endl;
+      DeleteFile(temp.c_str());
+    }
   }
-  if (deleteme) {
-    if (verbose) cout << "del \"" << temp << "\"" << endl;
-    DeleteFile(temp.c_str());
-  }
+}
+
+void tlib::Help(void) {
+  tool::Help();
+  string help = *archivearg.begin();
+  system(help.c_str());
 }
 
 void tlib::FoundFile(LI &i) {
@@ -36,6 +43,10 @@ void tlib::FoundFile(LI &i) {
 
 void tlib::FoundFlag(LI &i) {
   string temp = *i;
-  temp[0] = '/';
-  archivearg.push_back(temp);
+  if (temp == "-help") {
+    helpfound = -1;
+  } else {
+    temp[0] = '/';
+    archivearg.push_back(temp);
+  }
 }
